Rename directory and namespace
