+++ lib/CMakeLists.txt
@@ -452,4 +452,4 @@ add_subdirectory(CrashHandler)
 add_subdirectory(Assertions)
 add_subdirectory(Inspection)
 add_subdirectory(BuildId)
-add_subdirectory(Tasks)
+add_subdirectory(TaskMonitoring)
