+++ tests/CMakeLists.txt
@@ -424,4 +424,4 @@ add_subdirectory(AsyncRegistryServer)
 add_subdirectory(Containers)
 add_subdirectory(sepp)
 add_subdirectory(VocBase/Properties)
-add_subdirectory(Tasks)
+add_subdirectory(TaskMonitoring)
